# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Intuit exercise demo",
        "description": "OpenAPI documentation for APIs on Intuit's exercise demo",
        "contact": {
            "email": "alphazygma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Intuit example servers"
        }
    ],
    "paths": {
        "/project/{project_id}/bids": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "Get all Project's bids",
                "description": "Returns all bids from a given project",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bids"
                ],
                "summary": "Adds a bid on an existing project",
                "description": "Adds a bid to an existing project if valid and updates the project appropriately",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found|Invalid Input|Project is expired"
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of projects",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Creates a new project with the supplied definition, for the authenticated seller",
                "operationId": "store",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            }
        },
        "/projects/page/{page}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a list of projects",
                "description": "Returns list of projects for the supplied page",
                "operationId": "page",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page required for list of projects",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project information",
                "description": "Returns project data",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User information",
                "description": "Returns user data",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates an existing User",
                "description": "Updates an existing User data only for the elements suplied",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new User",
                "description": "Returns saved user data",
                "operationId": "store",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            }
        },
        "/sellers/{id}": {
            "get": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Get seller information",
                "description": "Returns seller data",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/sellers": {
            "post": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Add a new seller",
                "description": "Registers the current user as a seller",
                "operationId": "store",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            }
        },
        "/buyers/{id}": {
            "get": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Get buyer information",
                "description": "Returns buyer data",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/buyers": {
            "post": {
                "tags": [
                    "Buyers"
                ],
                "summary": "Add a new Buyer",
                "description": "Registers the current user as a buyer",
                "operationId": "store",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Input"
                    }
                }
            }
        }
    }
}